【front】Message系型をmedia/outputからmessage/{input,output}へ切り出し
