+++ frontend/src/api/internal/media/message.ts
@@ -1,6 +1,7 @@
 import { apiClient } from 'lib/axios/internal'
 import { ApiOut, apiOut } from 'lib/error'
-import { ChatMessage, MessageCreateIn, MessageUpdateIn } from 'types/internal/media/output'
+import { MessageCreateIn, MessageUpdateIn } from 'types/internal/message/input'
+import { ChatMessage } from 'types/internal/message/output'
 import { apiMessages, apiMessage } from 'api/uri'
 import { camelSnake } from 'utils/functions/convertCase'
