+++ frontend/src/components/templates/media/chat/detail.tsx
@@ -1,7 +1,9 @@
 import { useState, useRef, useEffect, useCallback, useMemo, FormEvent } from 'react'
 import { useRouter } from 'next/router'
 import { SubscribeIn } from 'types/internal/channel'
-import { ChatMessage, ChatReply, ChatDetailOut, MessageCreateIn, MessageUpdateIn } from 'types/internal/media/output'
+import { ChatDetailOut } from 'types/internal/media/output'
+import { MessageCreateIn, MessageUpdateIn } from 'types/internal/message/input'
+import { ChatMessage, ChatReply } from 'types/internal/message/output'
 import { LikeMediaIn } from 'types/internal/user'
 import { postSubscribeChannel } from 'api/internal/channel'
 import { postMessage, getReplies, putMessage, deleteMessage } from 'api/internal/media/message'
