+++ frontend/src/components/widgets/Message/SectionThread/index.tsx
@@ -1,5 +1,5 @@
 import { MouseEvent, RefObject } from 'react'
-import { ChatMessage, ChatReply } from 'types/internal/media/output'
+import { ChatMessage, ChatReply } from 'types/internal/message/output'
 import { UserMe } from 'types/internal/user'
 import cx from 'utils/functions/cx'
 import Divide from 'components/parts/Divide'
