+++ frontend/src/types/internal/media/output.d.ts
@@ -1,6 +1,6 @@
 import { Channel } from 'types/internal/channel'
 import { Comment } from 'types/internal/comment'
-import { Author } from 'types/internal/user'
+import { ChatMessage } from 'types/internal/message/output'
 
 export interface Search {
   name?: string
@@ -154,28 +154,4 @@ export interface BlogDetailOut {
 export interface ChatDetailOut {
   detail: ChatDetail
   list: Chat[]
-}
-
-export interface ChatMessage {
-  ulid: string
-  text: string
-  replyCount: number
-  created: Date
-  updated: Date
-  author: Author
-}
-
-export interface ChatReply extends ChatMessage {
-  parentUlid: string
-}
-
-export interface MessageCreateIn {
-  chatUlid: string
-  text: string
-  parentUlid?: string
-}
-
-export interface MessageUpdateIn {
-  chatUlid: string
-  text: string
 }
