+++ frontend/src/types/internal/message/input.d.ts
@@ -1,0 +1,10 @@
+export interface MessageCreateIn {
+  chatUlid: string
+  text: string
+  parentUlid?: string
+}
+
+export interface MessageUpdateIn {
+  chatUlid: string
+  text: string
+}
