+++ frontend/src/types/internal/message/output.d.ts
@@ -1,0 +1,14 @@
+import { Author } from 'types/internal/user'
+
+export interface ChatMessage {
+  ulid: string
+  text: string
+  replyCount: number
+  created: Date
+  updated: Date
+  author: Author
+}
+
+export interface ChatReply extends ChatMessage {
+  parentUlid: string
+}
